feat(retryWhen): add higher-order lettable version of retryWhen
